Refactor product details to work with individually fetched data
